release: 2.0.0.4-beta05
update: uhf sdk 2.0.0.8
fix(ui): restrict 255 settings

+++ app/src/main/java/com/gigatms/uhf/GeneralViewHolder.java
@@ -198,7 +198,7 @@ public class GeneralViewHolder extends RecyclerView.ViewHolder {
             public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
                 if (fromUser) {
                     textView.setText(String.valueOf(progress + seekBarParamData.getMinValue()));
-                    seekBarParamData.setSelected((byte) (progress + seekBarParamData.getMinValue()));
+                    seekBarParamData.setSelected(progress + seekBarParamData.getMinValue());
                 }
             }
 
@@ -352,7 +352,7 @@ public class GeneralViewHolder extends RecyclerView.ViewHolder {
             public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
                 if (fromUser) {
                     value.setText(String.valueOf(progress + seekBarTitleParamData.getMinValue()));
-                    seekBarTitleParamData.setSelected((byte) (progress + seekBarTitleParamData.getMinValue()));
+                    seekBarTitleParamData.setSelected(progress + seekBarTitleParamData.getMinValue());
                 }
             }
